requirements.txt
pyspark==3.0.1
jupyter==1.0.0
pandas
tensorflow-hub
tensorflow-gpu
keras
numpy
pytest
matplotlib
scikit-learn
pyyaml
ipykernel
kafka-python
findspark
torchtext==0.6.0
spacy
torch==1.7.0
GoogleNews==1.4.1
newspaper3k==0.2.8
pytz
searchtweets==1.7.6
xgboost==1.2.1
watermark==2.0.2
transformers==3.4.0
seaborn==0.11.0
vaderSentiment
pandas_datareader
pyarrow==2.0
dask[complete]==2.30.0
pyyaml
beautifulsoup4==4.9.3
selenium==3.141.0
retry==0.9.2